requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.151
botocore==1.34.151
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
diff-match-patch==20230430
dj-database-url==0.5.0
Django==5.0.7
django-allauth==0.63.5
django-chartjs==2.3.0
django-import-export==4.1.1
django-storages==1.14.4
et-xmlfile==1.1.0
gunicorn==22.0.0
jmespath==1.0.1
kombu==5.4.0
openpyxl==3.1.5
psycopg2==2.9.9
s3transfer==0.10.2
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
vine==5.1.0
